+++ src/main/java/Main.java
@@ -1,8 +1,12 @@
 public class Main {
     public static void main(String[] args) {
-        Calc calc = new Calc();
-        calc.getPeople();
-        calc.parseBill();
+        Check check = new Check();
+        // количество гостей
+        check.setGuests();
+        // позиции в чеке
+        check.addItems();
+        // результат
+        check.showResult();
     }
 
 }
